api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlmodel==0.0.22
SQLAlchemy==2.0.34
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.2
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
email-validator==2.2.0